refactor: auth controller and routes
